requirements.txt
google_api_python_client==2.177.0
matplotlib==3.10.6
pandas==2.3.2
python-dotenv==1.1.1
Requests==2.32.5
seaborn==0.13.2
selenium==4.35.0
streamlit==1.38.0
youtube_comment_downloader==0.1.78
